Fixes for collections and json parser. Added check for negative exponents and updated the parser to handle them correctly.
Multiple fixes for collections and additional error messages and clarification of how collections are managed and processed.

Collections are not correctly set so that they cannot be changed to non collection types as this would allow for potential loss of collected data. Collections can now be copied by assigning to another variable which creates a new collection.

A collection when assigned to an element of a collection will create a copy of the collection in the collection element.