Fix crash when collection elements are specified out of order var a = { 5:0, 1:1 }; The DSL allows creating of collection elements by specifying an index past any that are filled in, but the processing code was not handling the indexes correctly so DCS instruction failed with an exception. This fix also fixes issues where the element indexes are specified out of order in a non -sequential manner and issues where keys would return the incorrect element. Both situations were caused by element key index being incorrectly specified.
Fixes issue with function calls when used in expressions incorrectly calculating the expression. This was caused due to function call code being output directly instead of processed though the queue like the other instructions. The fix required that functions be handled like statements when they are invoked like a statement and like values when used in expressions.

Fix for disassembly with switch statement. The switch instruction JTB was skipping the first case block when disassembled this lead to confusing display of the switch IL codes.