break for while and for are working and fixed issues with if causing problems  with the statement block in which it is enclosed. This was due to using the same stack for if and the looping constructs. Fixed issue with switch when used with breaks in some positions could cause a crash when the switch was defined within a function. This was due to the output stack from the expression parser getting cleared due to exiting the expression parser early. This can occur when multiple levels of recursion are encountered. The fix was to move the tracking break list outside the expression function.
Fixed an issue with return +number; The lexer was writing the ADD token even though the value had earlier been determined to be a number. This happened in return because return looks ahead to see if the return statement has a value. It uses the PeekNextToken() method for this. But PeekNextToken() had not been updated to handle tracking previous and last read tokens. The fix was to restore the previous and last tokens after determining what the next token to be read is.

Refactored the PrintIssue() function so that there is now only a single PrintIssue(...) instead of one that takes variable parameters and one that does not.

Adding error checks for sequences where variables and values are placed in the program but no separators or operators are specified. For example, A A or 3 4. These sequences should be errors and showed up in function calls where it is more critical. For example, print(a b, c, d) should print a b c d but it was missing the b. Now an error is generated allowing the script to be fixed.